requirements.txt
async-timeout==4.0.2
bidict==0.22.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
Flask-SocketIO==5.3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
pymongo==4.3.3
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
redis==4.5.5
Werkzeug==2.3.4
